+++ data-migrations/process.sh
@@ -29,6 +29,7 @@ process_migration() {
   # See if the migration was already processed
   echo "Checking to see if $1 has been run ..."
   EXISTS=$(mariadb ${MYSQL_ARGS} -N ${MYSQL_DATABASE} <<< "SELECT * FROM dataMigrations WHERE migrationFile = '$1';")
+
   if [ -z "$EXISTS" ]; then
     # If not run it
     echo "NEW MIGRATION - $1. Processing migration ..."